move all config part from `game_state.rs` to `game_config.rs`
